Data Augmentation
Protein sequences of different organisms for the same purpose vary in length. This code takes the average size of data and either increases the length of sequence (if its lesser than average) by adding poly-A-tail or decreases the length (if its more than the average) by slicing the c and end terminals of the sequences.